Stub implementation of courses resource with test
Within the test I am defining a 'reference' structure that I expect all
`course` representations to use. The goal is to extract this out later
so that the reference and its test can be used across the various
services that will be generating or accepting this structure.

As part of test development we have a stubbed out implementation of this
structure in the courses controller