Extract contract testing
The contract testing has been moved out of this project and re-included
as a submodule.

The submodule is here: https://github.com/umn-asr/course_contract_tests

And it downloads the latest reference structure from here:
https://github.com/umn-asr/course_json_reference

The submodule version of the test caches the file, so it doesn't
download it every time. And it allows you to override the reference
structure with one of your own.