First pass at Repository
We are following the structure provided by this article:
https://medium.com/@KamilLelonek/why-is-your-rails-application-still-coupled-to-activerecord-efe34d657c91

The code in that article only supports finding data, though. I guess
saving data is left as an exercise for the reader. I have gotten
saving/updating working, though it may not be the best.

My biggest concern with this implementation is that the Repository
uses a lot of ActiveRecord-specific methods. And usage of orm_adapter is
currently leaking out of the more abstract Repository class into a
concrete CampusRepository class. I think this is fixable, but I wanted
to mention the problems.