Enforce abbreviation uniqueness at the db level
I don't see a way to validate uniqueness of an attribute in this current
repository approach.

Rails' `validates_uniqueness_of` is a validation that lives under
`ActiveRecord`, not under `ActiveModel`. So our `CampusEntity` does not
have access to it.

we could duplicate the behavior of this method at the Repository level,
I guess. We could also implement validations within `Campus`, which has
access to ActiveRecord.

But the whole point of separating out persistence from business logic is
to keep things from 'validations' away from your persistence layer,
right?

So, a problem to be solved.