Separate adapter from activerecord descendants
This commit is a first step towards adressing a problem I had with the
initial implementation.

In the implementation we followed, this class became both the
orm_adapter and the representation of the persisted data

```ruby
class Campus < ::ActiveRecord::Base
end
```

And I didn't like this. It meant when I dealt with `orm_adapter` inside
the repository I had to keep in mind that it was both a Campus and a
full copy of the ActiveRecord API. It also allowed me to take shortcuts.
Since I knew that it contained the entire AR API, I could write code
like

`orm_adapter.find(x).update_attributes(attrs)`

And doing so would mean that any other adapter we would later write
would have to implement all of these AR API methods.

For example, say we implement a OhmAdapter. In the previous repository,
the code for creating a new entity was:

```ruby
def create
  map_record_out(orm_adapter.new)
end
```

If we followed the existing approach, then the orm_adapter would be
`Storage::Models::Ohm::Campus` which would descend from `Ohm::Base`

And if you call `new` on an Ohm::Base class you get a
useless class back. You need to call `create`.

The previous implementation didn't give us a clear place to put that
sort of knowledge. But by separating the Adapter from the Persistince
Class, we now have a place. It would live in Adapters::Ohm

This code works, but needs further refactoring. As Davin pointed out,
there's never a situation where we'd be using the ActiveRecord Campus
but the Ohm Adapter. Right now the API allows that, but there's no good
reason to.