Render json using rabl
Since we will need to render both xml and json representation of the
data, we are switching to rabl since it will allow us to use one
template for both representations.  The attributes for the campus are
now getting pulled from the data store.  The objects for term and
courses are mocks so we can get tests to pass until we implement the
actual models.