Remove ReferenceTest check from controller
Fixes: https://github.com/umn-asr/courses/issues/8

Because of the submodule issue, and because of general design, we are
removing the controller's use of reference test.

Eventually, a service object will be here and it will encapsulate all of
this validation/rescue stuff. But it's not here yet. This code, while
not great, does get our tests to pass.