Change attr_accessor to attr_writer
We have a getter method for type, so attr_accessor is redundant.

type is not getting stored in the DB, we setting a default value in the
model.  Need the attr_writer so that calls to set the type can be
handled, although it is not necessary to set the type since it will
always be course.