Update search documentation
While I like the more concise search param names, it causes some
problems. Since a Course already has a subject, overriding that method
to allow searchability leads to the controller returning the overriden
subject instead of the subject resource.

So, I'm changing the search param names to use the more verbose
attribute names, which won't cause a problem when we override them.