extract JsonImport from db/seeds
JsonImport loads the contents of parsed json into active record
objects. This was a routine in db/seeds. It covered with an
integration-type test that loads the reference json and queries
the courses controller to make sure the values match (after doing
some formatting and sorting so rspec's deep hash compare can do
its thing)