Increase caching time to 2 days
Fixes issue #46

The import of data takes a long time (currently several hours). So our
plan is to rely on the cached data until we complete importing the new
data.

To support this, I'm greatly increasing the caching time to 48 hours.
Two days is more than we need, since we plan on importing new data once
a day. But I want to make sure that the cache will last until we
invalidate it after the import.