Unify presentation objects, force caching of everything
First off, this might not be the best solution. The problem I'm trying
to solve here is that our previous approach to caching database queries
was imperfect. For example, we cached the results of a query like.

```ruby
Course.include(:course_attributes).all.load
```

But when we then pulled that out of the cache and called
`course.course_attributes`, it would turn into a database hit. A very
slow database hit if someone were trying to find an attribute in a large
set of data.

Other attributes would have the same problem. From what I could tell, if
a has-and-belongs-to-many was involved, then caching wouldn't work.

A secondary problem I'm trying to solve here is that while we would
try to cache the query, we weren't then using that cached data in the
rabl templates. In the controller we were essentialy creating two
presentation objects: `SearchableCourses` and `CoursesPresenter`. The
result was that the rabl templates would then have to re-fetch a bunch
of database results, slowing them down even further. Eventhough all rabl
templates were being cached, it seemed dumb to make that initial hit
slower than it needed to be.

Thus, enter this collection of presenter objects. CourseController now
only creates one thing: CoursesPresenter. And it does it through the
declarative builder `fetch`. If this presenter exists in the cache, then
the cache is returned. Otherwise, it creates a CoursesPresenter and
caches it.

Creating a new new CoursesPresenter walks through all the data and
creates a tree of presenter objects. This forces active record to
resolve all its lazy Relation objects. So what we cache has all of the
data in it. I've tested this with all of our currently-supported search
methods and none of them trigger a DB hit.

The Query String Search methods we support are all defined at the
CoursePresenter level, since this is the top level object that Query
String Search will be using.

The goofy downside of this approach is that we have a tree of Presenter
objects that more or less mirrors our tree of ActiveRecord objects. If,
for example, we add a new attribute to the Section model, we will also
have to remember to add that attribute to the SectionPresenter.

There are ways around that, using the reflection methods given to us by
ActiveRecord. But implementing that now seemed a step too far. Mainly
because I don't think the attributes will change much. So doing a bunch
of reflection/meta-programming for something that won't changed seems
like cleverness for the sake of cleverness.

But if I'm wrong about that, and the attributes do end up changing a
bunch, we can revisit my approach.