alter directory import task for parallelization
change json_import:directory_import so that the courses get
imported in parallel. terms and campuses - because multiple files
could try to set the same one up - are still built in sequence.

Wrapping CoureJsonImport in
"ActiveRecord::Base.connection_pool.with_connection do" so that we
can get a db connection from the connection pool and know it will
be checked back in when the code completes.

Why Process.new and not Thread.new? Not sure, just trying processes
first and they seem to work so far.