Add `cache_key` methods to presenters
When rabl caches templates, it does so under the `cache_key` of the
resource it's showing. If cache_key isn't present, then it appears to be
using the Object ID (that's my guess, anyway).

If we have rabl show the same resource twice, the cached template is not
being used. What we need is a cache_key that won't change for the length
of time that resource exists.

So I'm adding cache_key methods to all our presenters. Rabl will call
these, get a static id value and then cache correctly. We tested this on
Staging and it was very effective.