sort json files to import by size
we are processing the json files in batches and the code is a bit
raw so we cannot start a new file until the batch completes, i.e. when
the biggest file completes. by sorting the files by size we at least can
process files that should take a similar amount of time together