Switch cache to redis
After experimenting with 3 caching strategies, I settled on Redis. The
three strategies I tried were:

- memcached/dalli
- memory_store
- redis

Memcached and Redis were about the same in terms of speed. Memory store
was much, much faster.

But memory store also had a pretty big drawback. After caching one
term's worth of data, the ruby process had consumed about 40% of the
server's memory. The server currently has 8 gig, and we can move to 16,
but it still seemed like we were going to run out of room pretty
quickly.

Also, memory store is wiped whenever passenger reboots. And the
cache-filling time is currently about 20-30 minutes per term. The idea
of having to go through that for every deploy was undesirable.

So, that left memcached and redis. Since their performance was the same,
I picked Redis for a few reasons:

- When we get to hot-swapping caches, this will be easier with Redis.
- Redis provides other features we can use in the app, such as easy
  tracking of the most-popular searches.
- Redis documentation is better.
- It's a solution we're already using for CSDS.