Cache courses individually, instead of as a group
This changes how we cache objects. We have been caching the full
CoursesPresenter object (and all of its sub-objects) as one big blob.
There are a couple problems with this.

- It's not optimal for key/value stores like Redis
- It introduces a performance roadblock.

For example, if a client were running a search that only returns one
course, our previous code would be forced to retrieve the entire
CoursesPresenter object. There is no way to just retrieve *some*
courses.

As is, this code will still retrieve the full CoursesPresenter object,
but it's the first step towards allowing us to only retrieve the courses
we care about.

To avoid db hits on Campus and Term, those are now cached separately.

And the value cached under "[campus_id]_[term_id]" is a collection of
all the cache keys for the courses underneath that campus & term.