Add 2-stage cache retrieval
In testing on our Staging environment, I found that our previous
approach to caching wasn't good enough for large data sets. The problem
(mostly) boiled down to us retrieving all courses from the cache before
doing the filtering.

With a course count around 8000 (which is what is offered in a term at
UMNTC), cache retrieval took about 25-30 seconds, using either Redis or
Memcached. With UMNDL (1700 courses), retrieval took about 5 seconds.

This time is largely wasted, since most of those courses will not end up
being shown. Users of this service will be filtering their results, so
we should only retrieve the courses that they actually want to see. Of
course, we still need enough data to perform the filtering.

To do this, CoursePresenter can now build a 'queryable' version of
itself. It has just enough data to respond to all of the filtering
methods that we support.

When caching, we now cache both the full version and the queryable
version of a course. They are stored under separate keys.

When a request comes in, we grab the queryable courses for the requested
campus/term. We then filter them using the query string. We then take
the remaining courses and retrieve the full versions from the cache.

I did some testing of this using a full term's worth of data, testing to
see how many seconds it took to return a single subject's worth of courses.

campus  old approach  new approach
umndl   5             1.2
umntc   25-30         3

Response time increases as you request more and more courses, obviously.
If someone requested 10 subjects worth of courses from UMNTC, that
response will be slower. But I think that sort of usage will be limited.

This commit fixes #54, more or less. We could still be faster.

Note that the current implementation of this code is terrible.
Refactoring to come.