Establish consistent cache_keys for models
Anything that descends from ActiveRecord will have a cache_key method.
It returns something like `[model_name]/[instance.id]`.

This is fine, but doesn't really work for us. For example, when we get a
request with the parameter `campus_id: umntc`, how do we retrieve that
Campus from the cache? If Campus is using its default cache_key, there
isn't a way.

In this commit, I'm establishing common cache_key getters and setters
for Course, Campus and Term. There are two versions of the getters: one
for the class, one for the instance. The class one seems weird, but it's
how we can do methods like

```
Campus.fetch('umntc')
```

The `cache_key_for_instance` method takes that 'umntc' param and returns
what the cache key for that campus instance would be. Then we can check
to see if that campus is cached.