Extract cache fetching/building logic from Presenters
A lot of logic and responsibilities had all gotten lumped together here.
Separating them.

QueryableCourses builds/retrieves a collection of QueryableCourse
objects. These are the top-level objects that will be used by
QueryStringSearch to determine which course's we'll eventually retrieve
and display.

CoursesPresenter and CoursePresenter are now (mostly) back to their main
roles of knowing how to present a collection of courses and a single
course, respectively.

CoursePresenter still has some responsibilities around caching. This is
because we don't cache Course objects, we cache CoursePresenter objects.
CoursePresenter objects are the ones that have resolved all the
ActiveRecord relations and are fully ready to display. If we cache
Course objects we end up with a bunch of unresolved Relation objects and
displaying the data requires a bunch of database access. See commit
c74fe3d4e432b47534b5348bc365bcf32f43edcf for a lot more detail on that.