add CachePool
CachePool has a pool of ActiveSupport::Cache::RedisStore objects.
It is designed to allow us to build up a cache (working with #next)
while we continue to serve requests out of a populated cache (working
with #current)

When the new cache has been built, it can become the current cache
by calling CachePool#next!