Turn on rabl cache_sources
Rabl documentation isn't great. This says that it "caches template
lookups" but doesn't really explain what that means. In my testing,
however, it does appear to speed up rendering of previously-rendered
courses. And it doesn't increase the amount of data in Redis, either.

The other caching option "cache_all_data" decreases performance and
should not be used.