Remove unnecessary RABL caching.
I thought that caching each resource was necessary, but it's currently
not.

We only display the sub-resources of Course as part of a course. And
Rabl currently caches all of the sub-resources as part of Course, thanks
to the cache command in the /courses/show.rabl view.

Removing the caching of the sub-resources has no effect on performance.
It does have the nice effect of making our Redis store a lot smaller.