Remove Rabl, use hash representations
While working on an idea for a cache improvement, I decided to look at
how rabl implemented caching. What I *thought* it did was cache the text
representation of the resource (i.e, a json string for a json response,
and an xml string for an xml response).

But, no, that's not what it does. It just builds a hash and caches it.
Then, when a request comes in, the hash is pulled out of the cache and
to_json or to_xml is called on it.

I played around with bypassing rabl entirely. If I pulled a rabl-created hash out of
the cache directly and called to_json|xml on it, the server response
looked exactly the same. And it was **much** faster.

And this was fine, but I was still using Rabl to generate these hashes
and stick them in a cache. Which meant that we could only populate the
cache by hitting actual URLs, which was going to cause some problems
with our cache rotation strategy.

There's nothing magical about how Rabl creates the cache. It's easily
replicated by giving all of our resources a `to_h` method and then
building & caching a hash representation of all of our courses, which is
what this commit does.

With the hashes in place, we no longer need any of the sub-resource
presenters. Or any of the rabl views. So those have been deleted.