Configure caching by environment
We have a lot of caching options that are being configured by
environment, and it's tricky to keep them all straight.

What this commit does is set a single flag in the environment config
file: config.caching.use

The cache initializer uses the value of that config to set up all the
caching correctly. If the value is set to true:

- rack caching on
- rails cache set to current CachePool instance
- FastCache (formerly MemoryCache) set to use MemoryCache.

If the value is set to false:

- rack caching off
- rails cache set to NullStore.
- FastCache set to NullStore

The default (set in application.rb) is to use caching. Test and
Development set it to false.