Simplify caching of QueryableCourse(s)
For some reason I made this more complex than it needs to be. I was
caching each queryable course individually, and then caching a
collection of all the queryable course ids. So reading them out of the
cache required two steps:

1. Retrieve the keys
2. Retrieve a queryable course for each key

I was probably doing this because of either a speed concern or because
of the idea that small cache entries are better than big cache entries.

But the speed of this approach is equivalent to just have one big cache
entry of queryable_courses. And since we always retrieve a collection of
queryable_courses, we might as well cache them as a collection. Also, it
makes the code way easier to parse.