Add MemoryCache, use it for QueryableCourses
Every request requires the server to pull all QueryableCourses out of
the cache. This can be slow for big campuses, since that set can be up
to 8000 items big.

So to make this part of the request as fast as possible, we can put
QueryableCourses into a memory cache. Retrieval out of Memory is much,
much faster.

But Memory caching has a drawback, every time we restart the passenger
process the cache is is cleared. So what I'm doing here is caching the
QueryableCourses collection in both the Memory and Redis cache. If,
after this caching, the memory cache is wiped, the Redis cache is
retrieved & put into the memory cache. This should be a nice bit of
redundancy.