Return 404 if url has invalid campus/term
Fixes #60

This fix is not my favorite, but I'm hard pressed to think of a reason
why we need something else. As this is the only place in the application
where we have to ask about the null-ness of the Campus & Term, we might
as well just do it in a conditional.