Convert CoursesPresenter to work with standard rendering
Terms#index and Campuses#index call to_h on their content before passing
into the rendering method. Courses didn't because CoursesPresenter
automagically returned a hash.

It makes more sense to explicitly call to_h so that we can see what is
going on. So I've converted CoursesPresenter.fetch to return an instance
of itself that handles the to_h call.