set base host for url helpers
RackCacheManager uses url helpers to figure out which urls to
pre-cache. The url helpers require that the host portion of the url
is configured, but production and staging did not have these set, so
the RackCacheManager blew up when it tried to generate the urls to
cache.

configure the host default_url_options for production, staging and
development. (Test was already configured.)