configure default url
RackCacheWarmer uses url heleprs which require the default_url hostname
is set.

Add an initializer to set the default_url_options :host. First it tries to
read the value from an environment variable, if it does not find anything,
it will use localhost:3000, the webrick (development/test) default.

This change depends on a parallel change to ansible which configures the
enviroment variables for production and staging see
(https://github.umn.edu/asrweb/asr_ansible/pull/40)