Require the presence of instruction_mode_id
Fixes #85

We see an occasional bug where a section in the XML file does not have an
instruction_mode. Instead of letting this section into the database, I
think we should reject it for being invalid.

If further research shows that it's valid for a section to not
have an instruction_mode, then we can revisit this fix.