fix json formatting issues highlighted by missing data
The test for importing json that contains a section with no grading
basis was failing even though the import code was fixed. It was
successfully importing the file but the returned json had
"grading_basis: {}", and we are expecting there to be no grading_basis
key.

Update the to_h method on course and section so the hash does not
contain a node for subresources (like a sections grading basis or
instruction mode) if it is empty. We will continue to show empty
attributes and empty arrays

after these changes the tests in course_json_import_spec pass