use a attribute mapping when parsing resources
instructor role is referred to as "role" in the reference json
but is stored as InstructorRole#abbreviation in the rails data model
(the json instructor resource is broken out into 2 data objects)
so we cannot use a simple array of attribute names to build it.

switch to using a hash that maps the json key to the rails attribute
name when parsing the resource from the json.