Revert schema to what comes from sqlite
The schema.rb file is altered depending on what db platform you run db
tasks against. Run a migration against Oracle, and a bunch of this file
changes. Run the migration against Sqlite afterwards and a bunch of
stuff changes back. This is a pain.

For the sake of simplicity and readability I'm reverting this back to
the Sqlite version. We should not check in changes made by running
migrations against Oracle.