Improve documentation.
As we moved a lot of Section resources into Course, it gave me a reason
to review the documentation and remember that it needed some work. This
commit fixes a few things.

First, I've added a courses.yml to describe the structure of the data
you get when visiting courses.json|xml.

Second, I've added some text to the Readme that points people to the
documentation.

Third, I've removed all of the incorrect descriptions that got
copy-and-pasted in to the documentation. When we can add real
descriptions, great. But until then I'd rather not have incorrect
descriptions.