Add ClassOffering model
When we changed the import to include data from the Course service, the
kind of data we ended up with in the database changed. It used to be we
only had Classes. But now we have Courses and Classes, and we need to
start being clearer about which is which.

In PeopleSoft terms, a Course is the description of a course that you
can take. It's the details of what the class will be about. But it does
not contain anything about the course actually being offered (locations,
instructors, sections, meeting_patterns, etc).

A Class is an actual offering of that Course. It contains the details
about who/when/where, etc.

In terms of our data model everything is a Course. Only some Courses
have sections for a campus/term and some don't. But we treat them
exactly the same.

What people want out of the courses.umn.edu service is actually Class
data. Clients want to search on things that are actually being offered.
This is what they had before we added Course data, so we need to
restore that ability.

We also need to retain a way to show all Courses (offered or not) so
that we can do things like create Course Catalogs.

This new ClassOffering (because of course we can't call it Class in
Ruby) model descends from Course and only changes the
`for_campus_and_term` scope so that it only returns records that have
sections.

I did not change any of the testing of Course's scope because all of the
test data that it uses has no sections, so we're already implicitly
testing that its scope returns sectionless records..