Add collection of QueryableClasses
Following the QueryableCourses/QueryableCourse pattern, I'm adding
similar files for Classes.

QueryableClass is an empty child of QueryableCourse. The behavior of the
two is identical, but I wanted to use a name that wouldn't lead to
confusion. I thought it would be odd if QueryableClasses created
QueryableCourse objects.

I've also changed the cache_key used by QueryableCourse so to avoid
collisions.