Use set conversion to remove duplicates
As the comment says, the scope returns 1 row per section, not one row
per course. I was using `distinct` to remove duplicates, but that fails
when used on an Oracle database because the Courses table has a CLOB
field.

The query that distinct generates is

```
SELECT DISTINCT COURSES.* FROM....
```

And Distinct and CLOBs do not get along.

There might be a way to resolve this with just ActiveRecord query
methods, but set conversion works just fine.