Do not import invalid Course json
Closes #111

There is a better way to do this, but it involves refactoring
course_json_import to use the Rails validations and we decided that
refactoring would happen at a later date.

For now, our fix is brute force. If the course_json is missing an
attribute that is required, we do not import that course.