Remove test for sections without grading basis
As of 40ae9c135 we expect that grading basis information is a resource
that belongs to a course. This test was written back when we thought
grading basis was associated with sections and we needed to handle the
case when the section had no grading basis information. Now that we have
fixed the data structure this test is no longer relevant.