Gemify cache_pool
The old implementation of CachePool had its configuration hardcoded.
This seemed necessary because we needed CachePool to have its
configuration loaded prior to the environment (e.g. staging.rb,
production.rb) getting loaded. The most common way to handle this
- initializers - do not work by default, because they get loadedi
after the environment files.

The hardcoding got the job done but once we wanted to change the
cache pool configuration, it was confusing to remember where to go
to change it. Since CachePool was written, I've learned more about
railties which allow us to change things in the rail loading process,
as long as they get loaded early enough in the rails startup.

Of two two options I came up with - requiring the railtie in
application.rb file or treating CachePool and its railtie like a gem,
I've chosen the latter. It requires a few more files, but it is a more
common patterni. Also the fewer customizations we have in
application.rb, the easier it is to upgrade rails (we can take
advantage of the rails update rake task and have less diffing to do).

This commit adds just enough (some folders, a gemspec and a Gemfile)
to make the CachePool and its railtie load like gem. I added tests for
the configuration behavior but have left them for now in the main
project's spec folder. The behavior of CachePool is still tied to rails
and testing them in the context of a rails application is easier. Plus
this 'gem' is not ready for prime time yet.