Document cache_pool_spec hackery
This spec uses an undocumented method Singleton.__init__ to set up
some of its tests. Documenting that method.

CachePool is a Singleton so it prevents us from initializing it.
However we want to test that it gets its settings from
Rails.configuration which requires us to initialize it several times. I
found a workaround in stackoverflow
http://stackoverflow.com/questions/1909181/how-to-test-a-singleton-class
that uses an undocumented method Singleton.__init__ to reload CachePool.