Update to Rails 5.0.0
Update Rails and other gems necessary to move to Rails 5.

Changes to app/controllers are to address deprecation warnings.

Changes to config files were done by the Rails upgrade task, with some
small tweaks to keep that task from destroying our custom configuration.

The test method that helped to normalize json needed to be changed as
the string values are now frozen. Simplest solution there was to replace
the frozen string with a downcased copy of the string. I also added a
new `normalize_json!` method that does all of the steps.

Though this app works as of this commit, there is one remaining config
option that we should update. I'll note that in an issue in the Repo.