Remove term_id from Courses table
This foreign key is no longer used. A course belongs to a subject which
belongs to a term.

I think this column is here from an earlier design. But it's currently
not populated. I think it's better to remove it than to periodically
wonder, "Hey, why are all the term_id values NULL? Something must be
wrong!"