Use Oracle for testing not SQLite
Removes test infrastructure that assumes we're using SQLite's in-memory
database for testing. Now we're testing against Oracle directly to
better replicate production.

Migrations were manually created previously and not tracked in `db/schema.rb`.
These migrations did not run against an empty database successfully and
have been removed.

With the move to Oracle as our test db, some of our tests broke due to
limitations in Oracle that weren't present when using SQLite. The tests
have been updated, as well as the actual database schema. The
`description` field on courses now has a 4000 char limit, in line with
Oracle's limits on varchar2 type fields and our needs for the test
(course descriptions can be quite long).