Update deploy config for new servers
Courses has been deployed to both of the new servers succesfully.

The DNS change hasn't happened yet, but since the Ruby version was
updated awhile ago, the repo has also not been deployable to the old
servers for awhile.

Updating the servers in the deploy config now in preparation for the DNS
change, and so that it is less to remember in every new branch.