Update Rails to 6.0.3.5 for CVE
Received an email notification for [CVE-2021-22880](GHSA-8hc4-xxm3-5ppp).

Since we do not use Postgres in this app, the CVE is not applicable and
so the update is not strictly necessary. But since we received the
warning and needed updates anyway for the new server (and other CVEs),
it seemed like a good time for an update regardless.

The update includes a new version of the autogenerated schema comments.

Tests still pass; deploy to the new server was successful.